Extracted image loader handler to a class.
The tests for DefaultImageLoader were failing intermittently. We think
this was due to using SDWebImageManager to load the image in the test,
so extracted our code (the callback function) to a new class and unit
tested that class.

We also moved setting the placeholder image to the table view controller
as to ensure that it shows while the image is loading.

Signed-off-by: Robert Gravina <rgravina@pivotal.io>