Remove all usages of UIWindow from tests.
Having tests use UIWindow was causing the test suite to fail when run on
CI. To avoid the problem, we decided to delete them as they provide only
limited value. It would be possible to add the router tests back by
using dependency injection, so we'll add that as a chore in Tracker.

Signed-off-by: Kyle Gearhart <kgearhart@pivotal.io>